+++ tests/test_package_api.py
@@ -53,6 +53,7 @@ def test_models_public_api() -> None:
     assert set(models.__all__) == {
         "CallendarVanDusenRTDModel",
         "IEC60751RTDModel",
+        "PolynomialRTDModel",
     }